LayoutTests/imported/w3c/web-platform-tests/css/cssom/CSSStyleSheet-constructable-disallow-import.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Illegal constructor

Harness Error (FAIL), message = Unhandled rejection: Illegal constructor

FAIL Inserting an @import rule through insertRule on a constructed stylesheet throws an exception assert_throws_dom: function "() => { (new CSSStyleSheet).insertRule(import_text) }" threw object "TypeError: Illegal constructor" that is not a DOMException SyntaxError: property "code" is equal to undefined, expected 12
TIMEOUT @import rules should not trigger any loads. Test timed out
FAIL @import rules are not parsed in CSSStyleSheet.replace Illegal constructor
FAIL @import rules are not parsed in CSSStyleSheet.replaceSync Illegal constructor

Harness Error (FAIL), message = Unhandled rejection: Illegal constructor

FAIL Inserting an @import rule through insertRule on a constructed stylesheet throws an exception assert_throws_dom: function "() => { (new CSSStyleSheet).insertRule(import_text) }" threw object "TypeError: Illegal constructor" that is not a DOMException SyntaxError: property "code" is equal to undefined, expected 12
TIMEOUT @import rules should not trigger any loads. Test timed out
FAIL @import rules are not parsed in CSSStyleSheet.replace Illegal constructor
FAIL @import rules are not parsed in CSSStyleSheet.replaceSync Illegal constructor
